connlib make IpProvider shared
